add category watches.
for now, how it works is: if you have a list of product categories (in a widget, for example), and you are logged in, then a <span> is inserted before the link.

the span has the class "product-categories-watch" and is initially invisible. if you want it to appear, give it some style. for example, "float:left;width:24px;height:24px;background:url(someicon.png)"